+++ yazi-actor/src/help/mod.rs
@@ -1,1 +1,1 @@
-yazi_macro::mod_flat!(arrow escape filter toggle);
+yazi_macro::mod_flat!(arrow close escape toggle);