+++ yazi-actor/src/lives/which.rs
@@ -24,10 +24,10 @@ impl Which {
 impl UserData for Which {
 	fn add_fields<F: UserDataFields<Self>>(fields: &mut F) {
 		fields.add_cached_field("tx", |_, me| Ok(me.tx.clone().map(yazi_binding::MpscUnboundedTx)));
-		fields.add_field_method_get("times", |_, me| Ok(me.inner.times));
 		fields.add_cached_field("cands", |lua, me| {
 			lua.create_sequence_from(me.inner.cands.iter().cloned())
 		});
+		fields.add_field_method_get("times", |_, me| Ok(me.inner.times));
 
 		fields.add_field_method_get("active", |_, me| Ok(me.inner.active));
 		fields.add_field_method_get("silent", |_, me| Ok(me.inner.silent));