+++ yazi-config/src/popup/mod.rs
@@ -1,1 +1,1 @@
-yazi_macro::mod_flat!(confirm input options pick);
+yazi_macro::mod_flat!(confirm help input options pick);