+++ yazi-core/src/help/help.rs
@@ -1,46 +1,33 @@
 use anyhow::Result;
-use unicode_width::UnicodeWidthStr;
+use ratatui_widgets::block::Padding;
+use yazi_binding::position::Position;
 use yazi_config::{KEYMAP, keymap::ChordArc};
-use yazi_macro::{act, render, render_and};
+use yazi_macro::render;
 use yazi_shared::Layer;
-use yazi_term::{TERM, event::{KeyCode, KeyEvent}};
+use yazi_term::event::KeyEvent;
 use yazi_tty::sequence::SetCursorStyle;
-use yazi_widgets::Scrollable;
-
-use crate::help::HELP_MARGIN;
+use yazi_widgets::{Scrollable, input::Input};
 
 #[derive(Default)]
 pub struct Help {
-	pub visible:         bool,
-	pub layer:           Layer,
-	pub(super) bindings: Vec<ChordArc>,
+	pub visible:  bool,
+	pub layer:    Layer,
+	pub position: Position,
+	pub bindings: Vec<ChordArc>,
 
 	// Filter
-	pub keyword:   String,
-	pub in_filter: Option<yazi_widgets::input::Input>,
+	pub input:   Input,
+	pub keyword: String,
 
 	pub offset: usize,
 	pub cursor: usize,
+	pub height: u16,
 }
 
 impl Help {
 	pub fn r#type(&mut self, key: KeyEvent) -> Result<bool> {
-		let Some(input) = &mut self.in_filter else { return Ok(false) };
-		match key {
-			KeyEvent { code: KeyCode::Escape, modifiers, .. } if modifiers.is_empty() => {
-				self.in_filter = None;
-				render!();
-			}
-			KeyEvent { code: KeyCode::Enter, modifiers, .. } if modifiers.is_empty() => {
-				self.in_filter = None;
-				return Ok(render_and!(true)); // Don't do the `filter_apply` below, since we already have the filtered results.
-			}
-			KeyEvent { code: KeyCode::Backspace, modifiers, .. } if modifiers.is_empty() => {
-				act!(backspace, input)?;
-			}
-			_ => {
-				input.r#type(key)?;
-			}
+		if !self.input.r#type(key)? {
+			return Ok(false);
 		}
 
 		self.filter_apply();
@@ -48,15 +35,20 @@ impl Help {
 	}
 
 	pub fn filter_apply(&mut self) {
-		let kw = self.in_filter.as_ref().map_or("", |i| i.value());
+		let kw = self.input.value();
 
 		if kw.is_empty() {
-			self.keyword = String::new();
+			self.keyword.clear();
 			self.bindings = KEYMAP.chords(self.layer).iter().cloned().collect();
 		} else if self.keyword != kw {
+			let lowercased = kw.to_lowercase();
 			self.keyword = kw.to_owned();
-			self.bindings =
-				KEYMAP.chords(self.layer).iter().filter(|&c| c.contains(kw)).cloned().collect();
+			self.bindings = KEYMAP
+				.chords(self.layer)
+				.iter()
+				.filter(|&c| c.desc_or_run().to_lowercase().contains(&lowercased))
+				.cloned()
+				.collect();
 		}
 
 		render!(self.scroll(0));
@@ -64,15 +56,7 @@ impl Help {
 }
 
 impl Help {
-	// --- Keyword
-	pub fn keyword(&self) -> Option<String> {
-		self
-			.in_filter
-			.as_ref()
-			.map(|i| i.value())
-			.or(Some(self.keyword.as_str()).filter(|&s| !s.is_empty()))
-			.map(|s| format!("Filter: {s}"))
-	}
+	pub fn padding(&self) -> Padding { Padding::new(1, 1, 1, 1) }
 
 	// --- Bindings
 	pub fn window(&self) -> &[ChordArc] {
@@ -81,27 +65,22 @@ impl Help {
 	}
 
 	// --- Cursor
-	pub fn cursor(&self) -> Option<(u16, u16)> {
-		if !self.visible || self.in_filter.is_none() {
-			return None;
-		}
-		if let Some(kw) = self.keyword() {
-			return Some((kw.width() as u16, TERM.dimension().rows));
-		}
-		None
-	}
+	pub fn cursor(&self) -> Option<u16> { self.visible.then_some(self.input.cursor()) }
 
 	pub fn rel_cursor(&self) -> usize { self.cursor - self.offset }
 
 	pub fn cursor_shape(&self) -> Option<SetCursorStyle> {
-		Some(self.in_filter.as_ref()?.cursor_shape())
+		self.visible.then_some(self.input.cursor_shape())
 	}
 }
 
 impl Scrollable for Help {
 	fn total(&self) -> usize { self.bindings.len() }
 
-	fn limit(&self) -> usize { TERM.dimension().rows.saturating_sub(HELP_MARGIN) as usize }
+	fn limit(&self) -> usize {
+		let p = self.padding();
+		self.height.saturating_sub(p.top + /* input */ 1 + /* divider */ 1 + p.bottom) as usize
+	}
 
 	fn cursor_mut(&mut self) -> &mut usize { &mut self.cursor }